+++ src/grandorgue/GOSetter.cpp
@@ -115,6 +115,7 @@ enum {
   ID_SETTER_CRESCENDO_B,
   ID_SETTER_CRESCENDO_C,
   ID_SETTER_CRESCENDO_D,
+  ID_SETTER_CRESCENDO_OVERRIDE,
 
   ID_SETTER_TEMPERAMENT_PREV,
   ID_SETTER_TEMPERAMENT_NEXT,
@@ -235,6 +236,7 @@ const struct ElementListEntry GOSetter::m_element_types[] = {
   {wxT("CrescendoPrev"), ID_SETTER_CRESCENDO_PREV, true, true},
   {wxT("CrescendoCurrent"), ID_SETTER_CRESCENDO_CURRENT, true, true},
   {wxT("CrescendoNext"), ID_SETTER_CRESCENDO_NEXT, true, true},
+  {wxT("CrescendoOverride"), ID_SETTER_CRESCENDO_OVERRIDE, true, true},
   {wxT(""), -1, false, false},
 };
 
@@ -300,6 +302,8 @@ GOSetter::GOSetter(GODefinitionFile *organfile)
 GOSetter::~GOSetter() {}
 
 void GOSetter::Load(GOConfigReader &cfg) {
+  m_organfile->RegisterSaveableObject(this);
+
   wxString buffer;
 
   m_framegeneral.resize(0);
@@ -319,9 +323,19 @@ void GOSetter::Load(GOConfigReader &cfg) {
   }
 
   m_crescendo.resize(0);
-  for (unsigned i = 0; i < 4 * CRESCENDO_STEPS; i++) {
+  for (unsigned i = 0; i < N_CRESCENDOS; i++) {
+    buffer.Printf(wxT("SetterCrescendo%d"), i);
+
+    bool defaultAddMode
+      = cfg.ReadBoolean(ODFSetting, buffer, wxT("AddMode"), false, false);
+
+    m_CrescendoAddMode[i] = cfg.ReadBoolean(
+      CMBSetting, buffer, wxT("AddMode"), false, defaultAddMode);
+  }
+  for (unsigned i = 0; i < N_CRESCENDOS * CRESCENDO_STEPS; i++) {
     m_crescendo.push_back(
       new GOFrameGeneral(m_organfile->GetGeneralTemplate(), m_organfile, true));
+    m_CrescendoExtraSets.emplace_back();
     buffer.Printf(
       wxT("SetterCrescendo%d_%03d"),
       (i / CRESCENDO_STEPS) + 1,
@@ -360,6 +374,10 @@ void GOSetter::Load(GOConfigReader &cfg) {
   m_button[ID_SETTER_CRESCENDO_B]->Init(cfg, wxT("SetterCrescendoB"), _("B"));
   m_button[ID_SETTER_CRESCENDO_C]->Init(cfg, wxT("SetterCrescendoC"), _("C"));
   m_button[ID_SETTER_CRESCENDO_D]->Init(cfg, wxT("SetterCrescendoD"), _("D"));
+  m_button[ID_SETTER_CRESCENDO_OVERRIDE]->Init(
+    cfg, wxT("SetterCrescendoOverride"), _("Override"));
+  m_button[ID_SETTER_CRESCENDO_OVERRIDE]->Display(
+    !m_CrescendoAddMode[m_crescendobank]);
 
   m_button[ID_SETTER_PITCH_M1]->Init(cfg, wxT("SetterPitchM1"), _("-1"));
   m_button[ID_SETTER_PITCH_M10]->Init(cfg, wxT("SetterPitchM10"), _("-10"));
@@ -407,6 +425,16 @@ void GOSetter::Load(GOConfigReader &cfg) {
     cfg, wxT("SetterGeneralPrev"), _("Prev"));
   m_button[ID_SETTER_GENERAL_NEXT]->Init(
     cfg, wxT("SetterGeneralNext"), _("Next"));
+}
+
+void GOSetter::Save(GOConfigWriter &cfg) {
+  for (unsigned i = 0; i < N_CRESCENDOS; i++) {
+    cfg.WriteBoolean(
+      wxString::Format(wxT("SetterCrescendo%d"), i),
+      wxT("AddMode"),
+      m_CrescendoAddMode[i]);
+  }
+  // another objects are saveble themself so they are saved separatelly
 }
 
 void GOSetter::ButtonChanged(int id) {
@@ -559,6 +587,15 @@ void GOSetter::ButtonChanged(int id) {
     m_crescendo[m_crescendopos + m_crescendobank * CRESCENDO_STEPS]->Push();
     break;
 
+  case ID_SETTER_CRESCENDO_OVERRIDE: {
+    GOButton *btn = m_button[ID_SETTER_CRESCENDO_OVERRIDE];
+    bool oldIsOverride = btn->IsEngaged();
+
+    m_CrescendoAddMode[m_crescendobank]
+      = oldIsOverride; // addMode == !isOverride
+    btn->Display(!oldIsOverride);
+  } break;
+
   case ID_SETTER_PITCH_M1:
     m_organfile->GetPipeConfig().ModifyTuning(-1);
     break;
@@ -670,6 +707,8 @@ void GOSetter::SetCrescendoType(unsigned no) {
   m_button[ID_SETTER_CRESCENDO_B]->Display(no == 1);
   m_button[ID_SETTER_CRESCENDO_C]->Display(no == 2);
   m_button[ID_SETTER_CRESCENDO_D]->Display(no == 3);
+  m_button[ID_SETTER_CRESCENDO_OVERRIDE]->Display(
+    !m_CrescendoAddMode[m_crescendobank]);
 }
 
 void GOSetter::ResetDisplay() {
@@ -723,12 +762,28 @@ void GOSetter::Crescendo(int newpos, bool force) {
   if (pos == m_crescendopos)
     return;
 
+  bool crescendoAddMode = m_CrescendoAddMode[m_crescendobank];
+
   while (pos > m_crescendopos) {
-    m_crescendo[++m_crescendopos + m_crescendobank * CRESCENDO_STEPS]->Push();
+    const unsigned oldIdx = m_crescendopos + m_crescendobank * CRESCENDO_STEPS;
+    const unsigned newIdx = oldIdx + 1;
+
+    if (crescendoAddMode)
+      m_crescendo[oldIdx]->GetExtraSetState(m_CrescendoExtraSets[oldIdx]);
+    else
+      m_CrescendoExtraSets[oldIdx].clear();
+    ++m_crescendopos;
+    m_crescendo[newIdx]->Push(
+      crescendoAddMode ? &m_CrescendoExtraSets[oldIdx] : nullptr);
   }
 
   while (pos < m_crescendopos) {
-    m_crescendo[--m_crescendopos + m_crescendobank * CRESCENDO_STEPS]->Push();
+    --m_crescendopos;
+
+    const unsigned newIdx = m_crescendopos + m_crescendobank * CRESCENDO_STEPS;
+
+    m_crescendo[newIdx]->Push(
+      crescendoAddMode ? &m_CrescendoExtraSets[newIdx] : nullptr);
   }
 
   wxString buffer;